chore: remove util.go
